Added typed variants of `value`.
`array` and `object`.

Both have const iterator support, which work the same as iterating
over a e.g. std::vector and std::map, respectively.

Also, subscript support for array an object values (also works for the
general `value` type).

Unfortunately, because of how C++ works, it's not possible to "hide"
the array indexing from the object type and vice-versa. At least, I'm
not motivated enough to solve it ;)

Also, a general `get_value()` was added, with optional default
value. For example, instead of using `get_integer_value()`, it's now
possible to get the same value by calling `get_value<int>()`.  The
default value is returned if the object is not of that specific type.